custom-elements/custom-element-lifecycle/types-of-callbacks/created-callback-invocation-test.html

<!DOCTYPE html>
<html>
<head>
<title>Created callback of a custom element should be is invoked after custom element instance is created and its definition is registered</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<meta name="assert" content="CREATED callback is invoked after custom element instance is created and its definition is registered">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#types-of-callbacks">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function() {
        callbackCalledCounter++;
    }
    var GeneratedConstructor = doc.registerElement('x-a', {prototype: proto});
    var customElement = new GeneratedConstructor();
    assert_equals(callbackCalledCounter, 1, 'Callback created should be called after ' +
        'element instance was created');
}, 'Test created callback when custom element is created by constructor');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function() {
        callbackCalledCounter++;
    }
    doc.registerElement('x-a-2', {prototype: proto});
    doc.body.innerHTML = '<x-a-2></x-a-2>';
    assert_equals(callbackCalledCounter, 1, 'Callback created should be called after ' +
        'element instance was created');
}, 'Test created callback when custom element is created in HTML');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function() {
        callbackCalledCounter++;
    }
    doc.registerElement('x-a-7', {prototype: proto});
    doc.body.innerHTML = '<div><x-a-7></x-a-7></div>';
    assert_equals(callbackCalledCounter, 1, 'Callback created should be called after ' +
        'element instance was created');
}, 'Test created callback when custom element is created in HTML as descendant of another element');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function() {
        callbackCalledCounter++;
    }
    doc.registerElement('x-a-3', {prototype: proto});
    var customElement = doc.createElement('x-a-3');
    assert_equals(callbackCalledCounter, 1, 'Callback created should be called after ' +
        'element instance was created');
}, 'Test created callback when custom element is created by createElement');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function() {
        callbackCalledCounter++;
    }
    doc.body.innerHTML = '<x-a-4></x-a-4>';
    doc.registerElement('x-a-4', {prototype: proto});
    assert_equals(callbackCalledCounter, 1, 'Callback created should be called after ' +
        'element instance was created');
}, 'Test created callback when custom element is created in HTML before ' +
    'registration of a custom element');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    doc.body.innerHTML = '<x-a-5 id="xa"></x-a-5>';
    var customElement = doc.querySelector('#xa');
    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function() {
        callbackCalledCounter++;
    }
    customElement.constructor.prototype = proto;
    assert_equals(callbackCalledCounter, 0,
        'Callback created should not be called if custom element is unresolved');
}, 'Test created callback if custom element is unresolved.');


testInIFrame('../../resources/x-element.html', function(doc) {
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function() {
        callbackCalledCounter++;
    }
    doc.registerElement('x-element', {prototype: proto});
    assert_equals(callbackCalledCounter, 1, 'Callback created should be called');
}, 'Test created callback. Create custom element, then register. ' +
    'Document has browsing context');


testInIFrame('../../resources/register-and-create-custom-element.html', function(doc) {
    assert_equals('Created callback was called', doc.querySelector('#log').innerText,
        'Callback created should be called')
}, 'Test created callback. Register custom element, then create. ' +
    'Document has browsing context');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function() {
        callbackCalledCounter++;
    }
    var customElement = doc.createElement('x-a-6');
    doc.registerElement('x-a-6', {prototype: proto});
    assert_equals(callbackCalledCounter, 1, 'Callback created should be called after ' +
        'registration of custom element');
}, 'Test created callback when custom element is created by createElement '+
    'before registration of a custom element');


test(function(){
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function() {
        callbackCalledCounter++;
    }
    var GeneratedConstructor = doc.registerElement('x-a-8', {prototype: proto});
    var customElement1 = new GeneratedConstructor();
    assert_equals(callbackCalledCounter, 1, 'Callback created should be called');

    var customElement2 = doc.createElement('x-a-8');
    assert_equals(callbackCalledCounter, 2, 'Callback created should be called after ' +
        'element instance was created');

    doc.body.innerHTML = '<x-a-8></x-a-8>';
    assert_equals(callbackCalledCounter, 3, 'Callback created should be called after ' +
        'element instance was created');

    doc.body.innerHTML = '<div><x-a-8></x-a-8></div>';
    assert_equals(callbackCalledCounter, 4, 'Callback created should be called after ' +
        'element instance was created');

}, 'Test created callback. Create several custom elements using different ways');


testInIFrame('../../resources/x-element.html', function(doc) {
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function() {
        callbackCalledCounter++;
    }
    var GeneratedConstructor = doc.registerElement('x-element', {prototype: proto});
    assert_equals(callbackCalledCounter, 1, 'Callback created should be called ' +
        'for custom element in loaded document');

    var customElement2 = new GeneratedConstructor();
    assert_equals(callbackCalledCounter, 2, 'Callback created should be called after ' +
        'element instance was created');

    var customElement3 = doc.createElement('x-element');
    assert_equals(callbackCalledCounter, 3, 'Callback created should be called after ' +
        'element instance was created');

    doc.body.innerHTML = '<x-element></x-element>';
    assert_equals(callbackCalledCounter, 4, 'Callback created should be called after ' +
        'element instance was created');

    doc.body.innerHTML = '<div><x-element></x-element></div>';
    assert_equals(callbackCalledCounter, 5, 'Callback created should be called after ' +
        'element instance was created');

}, 'Test created callback. Create several custom elements using different ways. ' +
    'Document has browsing context');

</script>
</body>
</html>
